Crate story_gen

This module contains most of the business logic required to run the application and for error handling.

Modules

grammar

This module contains all the grammar rule storage and generating random sentences. intiates the application.

random

Module that provides random helper functions that are used in the application.

Structs

Combinations

Used to encapsulate combinatorics logic for optional tokens.

Config

Validates the command line arguments, and stores their values.

Rule

Represents a grammar rule formed from a single line in the file provided.

Functions

convert_sentences_to_paragraphs

Converts generated sentences into paragraphs of given sentence length

parse_file

Convert lines from a file into grammar rules Program exits if any lines do not follow the rules listed in the README.md

parse_right_hand_side

Parses the string that that is to the rigth of the equal sign delimiter

process_rhs_optional_combinations

Converts the RHS vector into representing all possible combinations of optional tokens. If there are no optional tokens, then this method will return the same array.

run

Main function which runs and controls the life time of the application.